django_app/pyproject.toml
[project]
name = "django-telegram-auth"
version = "0.1.0"
description = "Simple app to authenticate users via Telegram"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.3.0",
    "django>=5.1.3",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.1"]

[tool.ruff.lint.mccabe]
max-complexity = 4

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D213"]

[tool.ruff]
exclude = ["**/.venv/**", "migrations"]
line-length = 120
